fix: streaming renderer 状态清理与 DSL 检测对齐
1. flush() 重置 current_block_type/current_block_index/current_acc，
   确保 MessageEnd 后渲染器回到初始状态
2. is_dsl_line() 从 DslParser 改为简单 :: 前缀检测，
   移除对 processor_chain::dsl_parser 的依赖
3. 补充 flush 状态清空和 DSL 检测的单元测试
4. 迁移 streaming 内联测试到独立文件 streaming_tests.rs

Source: CI
Type: fix
